Refactor env_resolver() into buck.config
Move it from buck.tox_hooks.tox_hook_3 as it is a common function for
manipulating configuration rather than interacting with the internals of
tox 3.
